+++ src/components/form-skeletons/hooks/useFormField.tsx
@@ -1,0 +1,66 @@
+import { getHintId, getErrorId, getErrorAriaLiveAttribute } from "../utils"
+import { ErrorValidationMode } from "../types"
+
+export type FormFieldData = {
+  controlProps: {
+    id: string
+    "aria-describedby": string | undefined
+    "aria-invalid": boolean
+    required: boolean
+  }
+  labelProps: {
+    htmlFor: string
+  }
+  hintProps: {
+    id: string
+    hidden: boolean
+  }
+  errorProps: {
+    id: string
+    hidden: boolean
+    "aria-live": `polite` | `assertive` | `off` | undefined
+  }
+}
+
+export function useFormField(
+  fieldId: string,
+  {
+    required = false,
+    hasError = false,
+    hasHint = false,
+    validationMode,
+  }: {
+    required?: boolean
+    hasError?: boolean
+    hasHint?: boolean
+    validationMode?: ErrorValidationMode
+  }
+): FormFieldData {
+  const hintId = getHintId(fieldId)
+  const errorId = getErrorId(fieldId)
+  const controlDescribedBy =
+    [hasError && errorId, hasHint && hintId]
+      .filter(describedBy => describedBy)
+      .join(` `) || undefined
+
+  return {
+    controlProps: {
+      id: fieldId,
+      "aria-describedby": controlDescribedBy,
+      "aria-invalid": hasError,
+      required,
+    },
+    labelProps: {
+      htmlFor: fieldId,
+    },
+    hintProps: {
+      id: hintId,
+      hidden: !hasHint,
+    },
+    errorProps: {
+      id: errorId,
+      hidden: !hasError,
+      "aria-live": getErrorAriaLiveAttribute(validationMode),
+    },
+  }
+}
