+++ src/components/form-skeletons/types.ts
@@ -1,0 +1,1 @@
+export type ErrorValidationMode = "focus" | "change" | "submit"
