+++ src/components/form-skeletons/utils/index.ts
@@ -1,3 +1,5 @@
+import { ErrorValidationMode } from "../types"
+
 export function getFinalAriaDescribedBy(
   controlDescribedBy?: string,
   ariaDescribedBy?: string
@@ -15,4 +17,16 @@ export function getHintId(fieldId: string) {
 
 export function getErrorId(fieldId: string) {
   return `${fieldId}__error`
+}
+
+export function getErrorAriaLiveAttribute(
+  validationMode?: ErrorValidationMode
+): React.HTMLAttributes<HTMLDivElement>["aria-live"] {
+  if (validationMode === `focus`) {
+    return `assertive`
+  }
+  if (validationMode === `change`) {
+    return `polite`
+  }
+  return undefined
 }
